gcp-terraform/app/bigquery/requirements.txt
functions-framework==3.*
google-cloud-storage==2.18
google-cloud-bigquery==3.*